project/backend/requirements.txt
Django==4.2.13
psycopg[binary]
djangorestframework
dj-database-url
gunicorn
requests